# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Operator360 Portal Backend API
  version: 1.0.0
  description: |
    Contract reference for the Operator360 portal backend, generated from the
    current handler implementations on branch f/oid-pid. All routes are under
    `/api` and require a bearer/session token validated by AuthMiddleware.

    Endpoints whose S3 file paths are keyed by an operator now resolve the S3
    prefix via `opt_master.data_path` looked up by `opt_id` — they no longer
    accept or require `opt_state` / `opt_district`.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
...
paths:
...
  /sid/batch_get:
    post:
      summary: Batch-fetch SID values from the upstream SID store
      operationId: GetSIDBatchValues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: array
                  items: { type: string }
                - type: object
                  required: [sids]
                  properties:
                    sids:
                      type: array
                      items: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  requested: { type: integer }
                  successful: { type: integer }
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        sid: { type: string }
                        success: { type: boolean }
                        status_code: { type: integer }
                        response: {}
                        error: { type: string }
        "400":
          description: Invalid body or empty sids
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500":
          description: Failed to load configuration
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...